Project setup task
Moves logic around installing dependencies and initial project setup from pre-commit to setup. This is done because .git is not commited with projects so it would not be shared between developers. `setup` is designed to be added to a project root and run the first time the project is cloned.